third_party/blink/web_tests/platform/generic/fast/dom/elementsFromPoint/elementsFromPoint-table-expected.txt

			
			
			
			
PASS nodeListToString(document.elementsFromPoint(125, 125)) is "TD#td11, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(275, 125)) is "TD#td14, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(175, 175)) is "TD#td22, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(125, 275)) is "TD#td41, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(100, 100)) is "TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(199, 199)) is "TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(125, 125)) is "TD#td14, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(275, 125)) is "TD#td11, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(100, 100)) is "TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(199, 199)) is "TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(125, 275)) is "TD#td14, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(275, 125)) is "TD#td41, TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(100, 100)) is "TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(199, 199)) is "TABLE#testtable, DIV#sandbox, BODY, HTML"
PASS successfullyParsed is true

TEST COMPLETE
